+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
+#### 0.9.0
+
+- feat: modify Dockerfile to pre-seed database during build - data is now baked into the Docker image
+
 #### 0.8.0
 
 - feat: update github action node version [`#2`](https://nova-eco/nova-eco/nova-db.git/pull/2)
@@ -10,6 +14,7 @@ All notable changes to this project will be documented in this file. Dates are d
 - feat: small typos and the addition of a README [`8c80aa1`](https://nova-eco/nova-eco/nova-db.git/commit/8c80aa144dbd78219d9c5da4b50d5d1bfa9c713e)
 - feat: update config with revised husky and commit lint files [`979ba5f`](https://nova-eco/nova-eco/nova-db.git/commit/979ba5ffb5160ad9e2108eeb06f8447c6b9e4039)
 - feat: update README [`fb84430`](https://nova-eco/nova-eco/nova-db.git/commit/fb84430df74d791fe2b2433bada9ed8099502d74)
+- fix: remove linting from the code quality workflow [`c32c9fc`](https://nova-eco/nova-eco/nova-db.git/commit/c32c9fc6a7b6596849cf11cd22746b959a628ae9)
 - feat: removed pnpm refs and replaced with npm [`a8a2403`](https://nova-eco/nova-eco/nova-db.git/commit/a8a24039dc8a7fcb66631a0284746c7450f2931c)
 - feat: add MIT license [`ab19e2d`](https://nova-eco/nova-eco/nova-db.git/commit/ab19e2d40ce106687a2c237df5fa4a9a6f4068a8)
 - feat: updated workflow [`c90b7fa`](https://nova-eco/nova-eco/nova-db.git/commit/c90b7faa5b28c731f0025d9737e1f0eae4679a2d)